+++ examples/dagster_examples_tests/test_toys/test_pandas_hello_world.py
@@ -5,12 +5,7 @@ from dagster_examples.toys.pandas_hello_world import (
     sum_solid,
 )
 
-from dagster import (
-    execute_pipeline,
-    execute_pipeline_with_preset,
-    execute_solid,
-    file_relative_path,
-)
+from dagster import execute_pipeline, execute_solid, file_relative_path
 
 
 def test_execute_pandas_hello_world_solid():
@@ -79,8 +74,8 @@ def test_execute_pandas_hello_world_pipeline_with_read_csv():
 
 
 def test_execute_hello_world_with_preset_test():
-    assert execute_pipeline_with_preset(pandas_hello_world_pipeline, 'test').success
+    assert execute_pipeline(pandas_hello_world_pipeline, preset='test').success
 
 
 def test_execute_hello_world_with_preset_prod():
-    assert execute_pipeline_with_preset(pandas_hello_world_pipeline, 'prod').success
+    assert execute_pipeline(pandas_hello_world_pipeline, preset='prod').success