+++ python_modules/dagster/dagster_tests/core_tests/execution_tests/test_run_config.py
@@ -1,4 +1,7 @@
+import re
 import uuid
+
+import pytest
 
 from dagster import PipelineDefinition, RunConfig, execute_pipeline, solid
 
@@ -37,6 +40,28 @@ def test_provided_run_id():
     assert called['yes']
 
 
+def test_select_mode():
+    called = {}
+
+    @solid
+    def check_run_id(context):
+        called['yes'] = True
+        assert uuid.UUID(context.run_id)
+        called['run_id'] = context.run_id
+
+    pipeline = PipelineDefinition(solid_defs=[check_run_id])
+
+    with pytest.warns(
+        UserWarning,
+        match=re.escape(
+            'In 0.8.0, the use of `run_config` to set pipeline mode and tags will be deprecated'
+        ),
+    ):
+        result = execute_pipeline(pipeline, run_config=RunConfig(mode='default'))
+    assert result.run_id == called['run_id']
+    assert called['yes']
+
+
 def test_injected_tags():
     called = {}
 
@@ -46,7 +71,13 @@ def test_injected_tags():
         called['yup'] = True
 
     pipeline_def = PipelineDefinition(name='injected_run_id', solid_defs=[check_tags])
-    result = execute_pipeline(pipeline_def, run_config=RunConfig(tags={'foo': 'bar'}))
+    with pytest.warns(
+        UserWarning,
+        match=re.escape(
+            'In 0.8.0, the use of `run_config` to set pipeline mode and tags will be deprecated'
+        ),
+    ):
+        result = execute_pipeline(pipeline_def, run_config=RunConfig(tags={'foo': 'bar'}))
 
     assert result.success
     assert called['yup']
@@ -74,5 +105,28 @@ def test_pipeline_tags():
     result = execute_pipeline(
         pipeline_def_with_override_tags, run_config=RunConfig(tags={'foo': 'bar'})
     )
+    assert result.success
+    assert called['yup']
+
+
+def test_overwrite_tags():
+    called = {}
+
+    @solid
+    def check_tags(context):
+        assert context.get_tag('foo') == 'baz'
+        called['yup'] = True
+
+    pipeline_def = PipelineDefinition(name='injected_run_id', solid_defs=[check_tags])
+    with pytest.warns(
+        UserWarning,
+        match=re.escape(
+            'In 0.8.0, the use of `run_config` to set pipeline mode and tags will be deprecated'
+        ),
+    ):
+        result = execute_pipeline(
+            pipeline_def, run_config=RunConfig(tags={'foo': 'bar'}), tags={'foo': 'baz'}
+        )
+
     assert result.success
     assert called['yup']