+++ python_modules/libraries/dagster-pyspark/dagster_pyspark_tests/test_decorators.py
@@ -1,14 +1,6 @@
 from dagster_pyspark import pyspark_resource, pyspark_solid
 
-from dagster import (
-    Field,
-    InputDefinition,
-    ModeDefinition,
-    execute_pipeline,
-    execute_pipeline_with_mode,
-    pipeline,
-    solid,
-)
+from dagster import Field, InputDefinition, ModeDefinition, execute_pipeline, pipeline, solid
 
 
 def test_simple_pyspark_decorator():
@@ -22,7 +14,7 @@ def test_simple_pyspark_decorator():
     def pipe():
         pyspark_job()
 
-    assert execute_pipeline_with_mode(pipeline=pipe, mode='default').success
+    assert execute_pipeline(pipeline=pipe, mode='default').success
 
 
 def test_pyspark_decorator_with_arguments():
@@ -40,7 +32,7 @@ def test_pyspark_decorator_with_arguments():
     def pipe():
         pyspark_job(produce_number())
 
-    assert execute_pipeline_with_mode(pipeline=pipe, mode='default').success
+    assert execute_pipeline(pipeline=pipe, mode='default').success
 
 
 def test_named_pyspark_decorator():
@@ -54,7 +46,7 @@ def test_named_pyspark_decorator():
     def pipe():
         pyspark_job()
 
-    assert execute_pipeline_with_mode(
+    assert execute_pipeline(
         pipeline=pipe,
         mode='default',
         environment_dict={'solids': {'blah': {'config': {'foo': 'baz'}}}},
@@ -90,7 +82,7 @@ def test_default_pyspark_decorator():
         first_pyspark_job()
         last_pyspark_job()
 
-    assert execute_pipeline_with_mode(pipeline=pipe, mode='default').success
+    assert execute_pipeline(pipeline=pipe, mode='default').success
 
 
 def test_aliased_pyspark_solid():